+++ simulation.cc
@@ -46,7 +46,7 @@ struct SimResult {
 //  * seed:    The seed to the random number generator.
 //  * print:   If true print out the results of the simulation.
 //  * returns: The success function and time it took to compute.
-SimResult working_set_simulator(CacheSim &sim, uint32_t seed, bool print = false) {
+SimResult working_set_simulator(CacheSim &sim, uint64_t seed, bool print = false) {
   std::mt19937_64 rand(seed);  // create random number generator
   auto start = absl::Now();
   for (uint64_t i = 0; i < kAccesses; i++) {
@@ -72,7 +72,7 @@ SimResult working_set_simulator(CacheSim &sim, uint32_t seed, bool print = false
   return {succ, duration};
 }
 
-SimResult uniform_simulator(CacheSim &sim, uint32_t seed, bool print = false) {
+SimResult uniform_simulator(CacheSim &sim, uint64_t seed, bool print = false) {
   std::mt19937_64 rand(seed); // create random number generator
   auto start = absl::Now();
   for (uint64_t i = 0; i < kAccesses; i++) {
@@ -106,7 +106,7 @@ SimResult simulate_on_seq(CacheSim &sim, std::vector<uint64_t>& seq, bool print 
   return {succ, duration};
 }
 
-std::vector<uint64_t> generate_zipf(uint32_t seed, double alpha) {
+std::vector<uint64_t> generate_zipf(uint64_t seed, double alpha) {
   // std::ofstream zipf_hist("Zipf_hist_" + std::to_string(alpha) + ".data");
   std::mt19937_64 rand(seed); // create random number generator
   std::vector<double> freq_vec;
